requirements.txt
requests
google-genai
yfinance
pydantic
python-dotenv
discord.py>=2.3.2
numpy
pandas
scikit-learn
pyarrow
duckdb
